requirements.txt
python-telegram-bot==13.15
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
pytz>=2023.3
apscheduler==3.6.3
beautifulsoup4>=4.12.0
lxml>=4.9.0
